[rs-admin] Fix Sonar issues.

+++ rs-admin/accessrights/accessrights-domain/src/main/java/fr/cnes/regards/modules/accessrights/domain/accountunlock/AccountUnlockToken.java
@@ -32,7 +32,7 @@ public class AccountUnlockToken {
     /**
      * Expiration delay in minutes (=24 hours)
      */
-    private static final long EXPIRATION = 60 * 24;
+    private static final int EXPIRATION = 60 * 24;
 
     /**
      * Id
